Change total stake seconds to u128

+++ programs/cardinal-stake-pool/src/state.rs
@@ -22,7 +22,7 @@ pub struct StakeEntry {
     pub original_mint_claimed: bool,
     pub last_staker: Pubkey,
     pub last_staked_at: i64,
-    pub total_stake_seconds: i64,
+    pub total_stake_seconds: i128,
     pub stake_mint_claimed: bool,
     pub stake_mint: Option<Pubkey>,
 }
